RLIB-16: Fix one more GCC 4.1 complaint about comparison between signed and unsigned. GCC 4.8 and GCC 5.3 can properly deduce there is no such problem because there is an if (n >= 0) just before the comparison GCC 4.1 complains about.

+++ libsrc/pcode_op_functions.c
@@ -1478,7 +1478,7 @@ gint rlib_pcode_operator_left(rlib *r, struct rlib_pcode *code, struct rlib_valu
 		gchar *tmp = g_strdup(RLIB_VALUE_GET_AS_STRING(v2));
 		gint n = RLIB_VALUE_GET_AS_NUMBER(v1)/RLIB_DECIMAL_PRECISION;
 		if (n >= 0) {
-			if (r_strlen(tmp) > n) *r_ptrfromindex(tmp, n) = '\0';
+			if (r_strlen(tmp) > (guint)n) *r_ptrfromindex(tmp, n) = '\0';
 		}
 		rlib_value_free(v1);
 		rlib_value_free(v2);
